Currently if you want to use inline bootstrap form rendering, this is usually enough:
1. Using bootstrap_3_layout.html.twig
2. Rendering form with `{{ form(form, { 'attr': {'class': 'form-inline'} }) }}`

Form rendering breaks for buttons as the buttons are rendered in a separate
<div> element without any attributes (this is coming from
form_div_layout.html.twig which has this block hard-coded). The rest of the
elements render with <div class="form-group"> which make them compatible with
form-inline attribute in the form (`<form class="form-inline">`).

The problem makes buttons render on new lines for inline forms. Extending
button_row in bootstrap_3_layout.html.twig template fixes this.

Signed-off-by: Edvinas Klovas <edvinas@pnd.io>